util: move zk.BigToFF and zk.BytesToArboStr to util package
this avoids an import cycle in upcoming refactor of circuit package
