+++ crypto/zk/utils_test.go
@@ -128,15 +128,3 @@ func TestLittleEndianToNBytes(t *testing.T) {
 	expected, _ = new(big.Int).SetString("873432238408170128747103711248787244651366455432", 10)
 	c.Assert(LittleEndianToNBytes(input, 20).Bytes(), qt.DeepEquals, expected.Bytes())
 }
-
-func TestBytesToArboStr(t *testing.T) {
-	c := qt.New(t)
-
-	input := new(big.Int).SetInt64(1233)
-	encoded := BytesToArboStr(input.Bytes())
-	expected := []string{
-		"145749485520268040037154566173721592631",
-		"243838562910071029186006881148627719363",
-	}
-	c.Assert(encoded, qt.DeepEquals, expected)
-}
